Refactoring: Consistent mapago layout
The idea of importing the "old" layout (only control plane)
into the new layout sucked.

The new layout is more consistent reflecting the
naming proposed by hgn.
