+++ control-plane/ctrl/client-protocols/udp-mcast.go
@@ -4,7 +4,7 @@ import "net"
 import "fmt"
 import "strconv"
 import "os"
-import "github.com/monfron/mapago/controlPlane/ctrl/shared"
+import "github.com/monfron/mapago/control-plane/ctrl/shared"
 import "io"
 
 const (
