+++ control-plane/ctrl/shared/common-funcs.go
